Rename:pencil:: Modify CoreDataManager filename
- 카테고리 내용 수정
